Bug fix: failed with required argparse arguments (#17)
Fixes the case that there are required arguments in your Snakefile
parser.  Other parts of your code will rely on the values from the
parser, but we don't care about those, so instead, we allow the
Snakefile compilation to fail, but only throw an exception if we are
unable to find the parsing code.  We don't care about the rest.